requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
decorator==4.4.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
proglog==0.1.10
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
setuptools==75.2.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
